treewide: Fix GCC compile warnings
Fix mostly trivial compile warnings emitted by GCC.

+++ google-modules/soc/gs/drivers/soc/google/cpif/link_rx_pktproc.c
@@ -374,7 +374,7 @@ static int pktproc_fill_data_addr(struct pktproc_queue *q)
 
 		if (unlikely(desc[fore].reserved0 != 0)) { /* W/A to detect mem poison */
 			mif_err("mem poison:0x%llX r0:%d c:%d s:%d l%d cl%d r1:%d\n",
-					desc[fore].cp_data_paddr, desc[fore].reserved0,
+					(u64)desc[fore].cp_data_paddr, desc[fore].reserved0,
 					desc[fore].control, desc[fore].status,
 					desc[fore].lro, desc[fore].clat, desc[fore].reserved1);
 			panic("memory poison\n");
